Move usecase implementations into domain/usecase package
