+++ app/usecase/get_scaffolds.go
@@ -8,18 +8,3 @@ import (
 type GetScaffoldsUseCase interface {
 	Perform() ([]scaffold.Scaffold, error)
 }
-
-type getScaffoldsUseCase struct {
-	repo scaffold.Repository
-}
-
-// NewGetScaffoldsUseCase creates a GetUseCase instance
-func NewGetScaffoldsUseCase(
-	repo scaffold.Repository,
-) GetScaffoldsUseCase {
-	return &getScaffoldsUseCase{repo: repo}
-}
-
-func (u *getScaffoldsUseCase) Perform() ([]scaffold.Scaffold, error) {
-	return u.repo.GetAll()
-}
