+++ domain/usecase/get_scaffolds.go
@@ -1,0 +1,21 @@
+package usecase
+
+import (
+	"github.com/izumin5210/scaffold/app/usecase"
+	"github.com/izumin5210/scaffold/domain/scaffold"
+)
+
+type getScaffoldsUseCase struct {
+	repo scaffold.Repository
+}
+
+// NewGetScaffoldsUseCase creates a GetUseCase instance
+func NewGetScaffoldsUseCase(
+	repo scaffold.Repository,
+) usecase.GetScaffoldsUseCase {
+	return &getScaffoldsUseCase{repo: repo}
+}
+
+func (u *getScaffoldsUseCase) Perform() ([]scaffold.Scaffold, error) {
+	return u.repo.GetAll()
+}
